# ucnrr_openapi.json
{"openapi":"3.1.0","info":{"title":"ReDNA UCN/RR Demo","version":"0.5.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/users/list":{"get":{"summary":"Users List","operationId":"users_list_users_list_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":500,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ingest_text":{"post":{"summary":"Ingest Text","description":"1) Save inbox text event (provenance).\n2) LLM two-pass extraction (explicit + implied).\n3) If still empty and USE_MIN_HEURISTICS=true, apply minimal fallback.\n4) If actionable, update rr_snapshot.json and call Core /ingest_from_ucnrr.","operationId":"ingest_text_ingest_text_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngestText"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/legacy_import":{"post":{"summary":"Legacy Import","operationId":"legacy_import_legacy_import_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LegacyImport"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IngestText":{"properties":{"user_id":{"type":"string","title":"User Id"},"text":{"type":"string","title":"Text"}},"type":"object","required":["user_id","text"],"title":"IngestText"},"LegacyImport":{"properties":{"user_id":{"type":"string","title":"User Id"},"legacy_key":{"type":"string","title":"Legacy Key"},"legacy_value":{"title":"Legacy Value"}},"type":"object","required":["user_id","legacy_key","legacy_value"],"title":"LegacyImport"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}